+++ docs/conf.py
@@ -64,7 +64,7 @@ author = u'David A. Riggs'
 # built documents.
 #
 # The short X.Y version.
-version = u'0'
+version = u'1.0'
 # The full version, including alpha/beta/rc tags.
 release = __version__